feat: Add Pokémon management commands and constants
- Introduced new constants for VS Code commands related to Pokémon management.
- Implemented commands for changing Pokémon language, configuring keybindings, exporting and importing Pokémon lists, deleting Pokémon, and spawning Pokémon.
- Created types for command dependencies and Pokémon specifications to streamline command implementations.
- Added functionality for managing Pokémon collections, including removing all Pokémon and performing roll calls.
- Enhanced user experience with quick pick options for selecting Pokémon and their attributes.
