+++ src/constants/vscode-keys.constant.ts
@@ -1,0 +1,29 @@
+// Built-in VS Code command keys
+export const VSCODE_SET_CONTEXT_COMMAND = 'setContext';
+
+// Extension context keys
+export const VSCODE_POKEMON_POSITION_CONTEXT_KEY = 'vscode-pokemon.position';
+
+// Extension view identifiers
+export const VSCODE_POKEMON_VIEW_ID = 'pokemonView';
+export const VSCODE_POKEMON_PANEL_VIEW_TYPE = 'pokemonCoding';
+
+// Extension command identifiers
+export const VSCODE_POKEMON_START_COMMAND = 'vscode-pokemon.start';
+export const VSCODE_POKEMON_SPAWN_POKEMON_COMMAND =
+  'vscode-pokemon.spawn-pokemon';
+export const VSCODE_POKEMON_SPAWN_RANDOM_POKEMON_COMMAND =
+  'vscode-pokemon.spawn-random-pokemon';
+export const VSCODE_POKEMON_DELETE_POKEMON_COMMAND =
+  'vscode-pokemon.delete-pokemon';
+export const VSCODE_POKEMON_REMOVE_ALL_POKEMON_COMMAND =
+  'vscode-pokemon.remove-all-pokemon';
+export const VSCODE_POKEMON_ROLL_CALL_COMMAND = 'vscode-pokemon.roll-call';
+export const VSCODE_POKEMON_CONFIGURE_KEYBINDINGS_COMMAND =
+  'vscode-pokemon.configure-keybindings';
+export const VSCODE_POKEMON_CHANGE_LANGUAGE_COMMAND =
+  'vscode-pokemon.change-pokemon-language';
+export const VSCODE_POKEMON_EXPORT_LIST_COMMAND =
+  'vscode-pokemon.export-pokemon-list';
+export const VSCODE_POKEMON_IMPORT_LIST_COMMAND =
+  'vscode-pokemon.import-pokemon-list';
