+++ src/extension/commands/types/pokemon-panel.type.ts
@@ -1,0 +1,9 @@
+import { PokemonSpecificationLike } from './pokemon-specification.type';
+
+export interface PokemonPanelLike {
+  resetPokemon(): void;
+  spawnPokemon(spec: PokemonSpecificationLike): void;
+  deletePokemon(pokemonName: string): void;
+  listPokemon(): void;
+  rollCall(): void;
+}
